refactor(queue): colocate consumer interfaces
